knowledge_base/pyproject.toml
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chatwilly-knowledge"
version = "0.1.0"
description = "Backend Knowledge"
readme = "README.md"
requires-python = ">=3.13"
authors =[
    { name = "William Simoni" }
]
dependencies =[
    "pydantic==2.12.5",
    "langchain==1.2.10",
    "langchain-openai==1.1.10",
    "langchain-core==1.2.16",
    "langgraph==1.0.10",
    "PyYAML==6.0.3",
    "pydantic-settings==2.13.1",
    "pypdf==6.7.5",
    "python-docx==1.2.0"
]

[project.optional-dependencies]
dev =[
    "pytest",
    "ruff",
    "mypy",
    "pre-commit"
]

[tool.setuptools.packages.find]
where = ["src"]
include =["chatwilly_backend*"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = [
    "src",
]
testpaths = [
    "tests",
]
